requirements.txt
# Core
python-dotenv==1.0.1
httpx==0.27.2
tenacity==8.5.0
structlog==24.4.0
asyncio-throttle==1.0.2

# Supabase
supabase==2.7.4
vecs==0.4.3

# Claude Agent SDK
claude-agent-sdk==0.1.51
anthropic==0.34.2

# Telegram
python-telegram-bot==21.5

# Amazon SP-API auth
boto3==1.35.23
botocore==1.35.23
requests==2.32.3

# Scheduling
schedule==1.2.2

# Utilities
pydantic==2.8.2
pydantic-settings==2.4.0
python-dateutil==2.9.0
pytz==2024.1
aiofiles==24.1.0

# Dev
pytest==8.3.2
pytest-asyncio==0.23.8
